feat: slashing reporting (#604)
## Description

The introduced method acts a safety net for `submitWithdrawals` method.
A proof of validator slashing should be delivered first to be able to
report `slashingPenalty`.

A separate role, `SUBMIT_WITHDRAWALS_ROLE`, was introduced to
distinguish between different actors. This role will be granted to both
`CSVerifier` and `EvmScriptExecutor`, while the `VERIFIER_ROLE` will be
granted exclusively to `CSVerifier`.

## Checklist

- [x] Appropriate PR labels applied
- [x] Test coverage maintained (`just coverage`)
  - [ ] No need to add/update tests
  - [x] Tests are added/updated
- [x] Documentation maintained
  - [ ] No need to update
  - [x] Updated